+++ packages/restapi/src/lib/chat/chats.ts
@@ -1,8 +1,8 @@
 import axios from 'axios';
-import { getAPIBaseUrls, isValidETHAddress, walletToPCAIP10 } from '../helpers';
-import Constants, {ENV} from '../constants';
+import { getAPIBaseUrls, isValidETHAddress } from '../helpers';
+import Constants, { ENV } from '../constants';
 import { IFeeds } from '../types';
-import { getInboxLists } from './helpers';
+import { getInboxLists, getUserDID, addDeprecatedInfo } from './helpers';
 
 export type ChatsOptionsType = {
   account: string;
@@ -12,29 +12,45 @@ export type ChatsOptionsType = {
    */
   toDecrypt?: boolean;
   /**
+   * page index - default 1
+   */
+  page?: number;
+  /**
+   * no of items per page - default 10 - max 30
+   */
+  limit?: number;
+  /**
    * Environment variable
    */
-  env?: ENV ;
+  env?: ENV;
 };
 
 /**
  * Return the latest message from all wallet addresses you have talked to. This can be used when building the inbox page.
  */
 export const chats = async (options: ChatsOptionsType): Promise<IFeeds[]> => {
-  const { account, pgpPrivateKey, env = Constants.ENV.PROD, toDecrypt = false } = options || {};
-  const user = walletToPCAIP10(account);
+  const {
+    account,
+    pgpPrivateKey,
+    env = Constants.ENV.PROD,
+    toDecrypt = false,
+    page = 1,
+    limit = 10,
+  } = options || {};
+  if (!isValidETHAddress(account)) {
+    throw new Error(`Invalid address!`);
+  }
+  const user = await getUserDID(account, env);
   const API_BASE_URL = getAPIBaseUrls(env);
-  const apiEndpoint = `${API_BASE_URL}/v1/chat/users/${user}/chats`;
+  const apiEndpoint = `${API_BASE_URL}/v1/chat/users/${user}/chats?page=${page}&limit=${limit}`;
   const requestUrl = `${apiEndpoint}`;
   try {
-    if (!isValidETHAddress(user)) {
-      throw new Error(`Invalid address!`);
-    }
     const response = await axios.get(requestUrl);
     const chats: IFeeds[] = response.data.chats;
+    const updatedChats = addDeprecatedInfo(chats);
     const feeds: IFeeds[] = await getInboxLists({
-      lists: chats,
-      user,
+      lists: updatedChats,
+      user: user,
       toDecrypt,
       pgpPrivateKey,
       env,