+++ packages/restapi/src/lib/user/createUser.ts
@@ -1,6 +1,7 @@
 import {
   createUserService,
   generateKeyPair,
+  generateRandomSecret,
   getAccountAddress,
   getWallet,
 } from '../chat/helpers';
@@ -10,12 +11,15 @@ import {
   walletToPCAIP10,
   encryptPGPKey,
   preparePGPPublicKey,
+  isValidCAIP10NFTAddress,
+  validatePssword,
 } from '../helpers';
 import {
   SignerType,
   encryptedPrivateKeyType,
   ProgressHookType,
   IUser,
+  encryptedPrivateKeyTypeV2,
 } from '../types';
 
 export type CreateUserProps = {
@@ -23,15 +27,26 @@ export type CreateUserProps = {
   account?: string;
   signer?: SignerType;
   version?: typeof Constants.ENC_TYPE_V1 | typeof Constants.ENC_TYPE_V3;
+  additionalMeta?: {
+    NFTPGP_V1?: {
+      password: string;
+    };
+  };
   progressHook?: (progress: ProgressHookType) => void;
 };
 
 export const create = async (options: CreateUserProps): Promise<IUser> => {
+  const passPrefix = '$0Pc'; //password prefix to ensure password validation
   const {
     env = Constants.ENV.PROD,
     account = null,
     signer = null,
     version = Constants.ENC_TYPE_V3,
+    additionalMeta = {
+      NFTPGP_V1: {
+        password: passPrefix + generateRandomSecret(10),
+      },
+    },
     progressHook,
   } = options || {};
 
@@ -46,12 +61,20 @@ export const create = async (options: CreateUserProps): Promise<IUser> => {
     if (!isValidETHAddress(address)) {
       throw new Error(`Invalid address!`);
     }
+    if (additionalMeta?.NFTPGP_V1?.password) {
+      validatePssword(additionalMeta.NFTPGP_V1.password);
+    }
 
     const caip10: string = walletToPCAIP10(address);
     let encryptionType = version;
 
-    // falback to v1
-    if (!signer) encryptionType = Constants.ENC_TYPE_V1;
+    if (isValidCAIP10NFTAddress(caip10)) {
+      // upgrade to v4 (nft encryption)
+      encryptionType = Constants.ENC_TYPE_V4;
+    } else {
+      // downgrade to v1
+      if (!signer) encryptionType = Constants.ENC_TYPE_V1;
+    }
 
     // Report Progress
     progressHook?.({
@@ -85,12 +108,23 @@ export const create = async (options: CreateUserProps): Promise<IUser> => {
         'Encrypting your keys. Please sign the message to continue.',
       level: 'INFO',
     });
+
     const encryptedPrivateKey: encryptedPrivateKeyType = await encryptPGPKey(
       encryptionType,
       keyPairs.privateKeyArmored,
-      address,
-      wallet
+      wallet,
+      additionalMeta
     );
+
+    if (encryptionType === Constants.ENC_TYPE_V4) {
+      const encryptedPassword: encryptedPrivateKeyTypeV2 = await encryptPGPKey(
+        Constants.ENC_TYPE_V3,
+        additionalMeta.NFTPGP_V1?.password as string,
+        wallet,
+        additionalMeta
+      );
+      encryptedPrivateKey.encryptedPassword = encryptedPassword;
+    }
 
     const body = {
       user: caip10,
@@ -99,8 +133,10 @@ export const create = async (options: CreateUserProps): Promise<IUser> => {
       encryptedPrivateKey: JSON.stringify(encryptedPrivateKey),
       encryptionType: encryptionType,
       env,
-      encryptedPassword: null,
-      nftOwner: null,
+      nftOwner:
+        encryptionType === Constants.ENC_TYPE_V4
+          ? walletToPCAIP10((await signer?.getAddress()) as string)
+          : null, // check for nft
     };
 
     // Report Progress