+++ packages/socket/src/lib/config.ts
@@ -3,8 +3,12 @@ type APIUrlType = {
     [key: string]: string
 };
 
-export const API_URLS : APIUrlType = {
+export const API_URLS: APIUrlType = {
   [ENV.PROD]: 'https://backend.epns.io',
   [ENV.STAGING]: 'https://backend-staging.epns.io',
-  [ENV.DEV]: 'https://backend-dev.epns.io'
+  [ENV.DEV]: 'https://backend-dev.epns.io',
+  /**
+   * **This is for local development only**
+   */
+  [ENV.LOCAL]: 'http://localhost:4000',
 };